% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{add_design}
\alias{add_design}
\title{Add two designs}
\usage{
add_design(design_1, design_2, alpha)
}
\arguments{
\item{design_1}{A dataframe with 'Point' and 'Weight' as column that represent the first design to add}

\item{design_2}{A dataframe with 'Point' and 'Weight' as column that represent the second design to add}

\item{alpha}{Weight of the first design}
}
\value{
A design as a dataframe with the weighted addition of the two designs
}
\description{
Add two designs
}
