% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{delta_bound}
\alias{delta_bound}
\title{Calculate efficiency bounds}
\usage{
delta_bound(alpha, k, sens_min, sens_max = Inf)
}
\arguments{
\item{alpha}{Combined weight of the new points to add}

\item{k}{Number of parameters of the model}

\item{sens_min}{Minimum value of the sensitivity function}

\item{sens_max}{Maximum value of the sensitivity function}
}
\value{
A numeric vector with two components, minimum and maximum efficiency (over 1)
}
\description{
Given the weight of new points, number of parameters and range of the sensitivity
function, calculates the range of possible minimum efficiency for the D-augmented design.
}
\examples{
optedr:::delta_bound(0.3, 3, 1.3, 3)
optedr:::delta_bound(1/4, 3, -2.50965)
}
