% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_mat_sens.R
\name{dsens}
\alias{dsens}
\title{Sensitivity function for D-Optimality}
\usage{
dsens(grad, M)
}
\arguments{
\item{grad}{A function in a single variable that returns the partial derivatives vector of the model.}

\item{M}{Information Matrix for the sensitivity function.}
}
\value{
The sensitivity function as a matrix of single variable.
}
\description{
Calculates the sensitivity function from the gradient vector and the Identity Matrix.
}
