% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_eff.R
\name{efficiency_des}
\alias{efficiency_des}
\title{Efficiency between optimal design and a user given design}
\usage{
efficiency_des(opt_des_obj, design)
}
\arguments{
\item{opt_des_obj}{An object given by the function \code{opt_des}.}

\item{design}{A dataframe that represents the design. Must have two columns:
\itemize{
\item \code{Point} contains the support points of the design.
\item \code{Weight} contains the corresponding weights of the \code{Point}s.
}}
}
\value{
The efficiency as a value between 0 and 1
}
\description{
Takes a optimal design provided from the function \code{opt_des} and a user given design and compares their
efficiency
}
\examples{
result <- opt_des("D-Optimality", y ~ a*exp(-b/x), c("a", "b"), c(1, 1500), c(212, 422))
design <- data.frame("Point" = c(220, 240, 400), "Weight" = c(1/3, 1/3, 1/3))
efficiency(result, design)
}
\seealso{
opt_des
}
