% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCross2}
\alias{getCross2}
\title{Give effective limits to candidate points region}
\usage{
getCross2(cross, min, max, start, par)
}
\arguments{
\item{cross}{Vector of crosspoints in the sensitivity function given an efficiency and weight}

\item{min}{Mininum of the space of the design}

\item{max}{Maximum of the space of the design}

\item{start}{Boolean that gives the effective start of the candidate points region}

\item{par}{Boolean with the parity of the region}
}
\value{
Vector of effective limits of the candidate points region. Taken
in pairs from the beginning delimit the region.
}
\description{
Given the start of the candidates points region, the parity of the crosspoints
and the boundaries of the space of the design returns the effective limits of
the candidate points region. Those points, taken in pairs from the first to
the last delimit the region.
}
\examples{
optedr:::getCross2(c(3, 5, 7), -5, 10, TRUE, FALSE)
}
