% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStart}
\alias{getStart}
\title{Find where the candidate points region starts}
\usage{
getStart(cross, min, max, val, sens_opt)
}
\arguments{
\item{cross}{Vector of crosspoints in the sensitivity function given an efficiency and weight}

\item{min}{Minimum of the space of the design}

\item{max}{Maximum of the space of the design}

\item{val}{Value of the sensitivity function at the crosspoints}

\item{sens_opt}{Sensitivity function}
}
\value{
True if the candidate points region starts on the minimum, False otherwise
}
\description{
Given the crosspoints and the sensitivity function, this function
finds where the candidate points region starts, either on the extreme of
the space of the design or the first crosspoints
}
