% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_mat_sens.R
\name{inf_mat}
\alias{inf_mat}
\title{Information Matrix}
\usage{
inf_mat(grad, design)
}
\arguments{
\item{grad}{A function in a single variable that returns the partial derivatives vector of the model.}

\item{design}{A dataframe that represents the design. Must have two columns:
\itemize{
\item \code{Point} contains the support points of the design.
\item \code{Weight} contains the corresponding weights of the \code{Point}s.
}}
}
\value{
The information matrix of the design, a \eqn{k\times k} matrix where k is the length of the gradient.
}
\description{
Given the gradient vector of a model in a single variable model and a design, calculates the information matrix.
}
