% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integrate_reg_int}
\alias{integrate_reg_int}
\title{Integrate IM}
\usage{
integrate_reg_int(grad, k, reg_int)
}
\arguments{
\item{grad}{function of partial derivatives of the model.}

\item{k}{number of unknown parameters of the model.}

\item{reg_int}{optional numeric vector of two components with the bounds of the interest region for I-Optimality.}
}
\value{
The integrated information matrix.
}
\description{
Integrates the information matrix over the region of interest to calculate matrix B to be used in I-Optimality
calculation.
}
