% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_mult.R
\name{IWFMult}
\alias{IWFMult}
\title{Cocktail Algorithm implementation for L-, I- and A-Optimality}
\usage{
IWFMult(
  init_design,
  grad,
  matB,
  design_space,
  grid.length,
  join_thresh,
  delete_thresh,
  delta_weights,
  tol,
  tol2,
  criterion,
  max_iter
)
}
\arguments{
\item{init_design}{optional dataframe with the initial design for the algorithm.}

\item{grad}{function of partial derivatives of the model.}

\item{matB}{optional matrix for L-optimality.}

\item{design_space}{named list with bounds for each design variable.}

\item{grid.length}{numeric value for the sensitivity search grid / LHS size.}

\item{join_thresh}{numeric value for the merge heuristic.}

\item{delete_thresh}{numeric value for the weight deletion threshold.}

\item{delta_weights}{numeric value in (0, 1), parameter of the algorithm.}

\item{tol}{numeric value for convergence of the weight loop.}

\item{tol2}{numeric value for the outer stop condition.}

\item{criterion}{character variable with the chosen optimality criterion.}

\item{max_iter}{maximum number of outer iterations.}
}
\value{
An object of class \code{optdes}.
}
\description{
Cocktail Algorithm implementation for L-, I- and A-Optimality
}
\seealso{
Other cocktail algorithms: 
\code{\link{CWFMult}()},
\code{\link{DWFMult}()},
\code{\link{DsWFMult}()},
\code{\link{WFMult}()}
}
\concept{cocktail algorithms}
