% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot Convergence of the algorithm}
\usage{
plot_convergence(convergence)
}
\arguments{
\item{convergence}{A dataframe with two columns:
\itemize{
\item \code{criteria} contains value of the criterion on each step.
\item \code{step} contains number of the step.
}}
}
\value{
A ggplot object with the \code{criteria} in the \code{y} axis and \code{step} in the \code{x} axis.
}
\description{
Plots the criterion value on each of the steps of the algorithm, both for optimizing weights and points,
against the total step number.
}
\examples{
conv <- data.frame("criteria" = c(24, 23, 20, 15, 14.9, 14.8, 14.7), "step" = 1:7)
optedr:::plot_convergence(conv)
}
