% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{print.augment_region}
\alias{print.augment_region}
\title{Print method for augment_region objects}
\usage{
\method{print}{augment_region}(x, ...)
}
\arguments{
\item{x}{An object of class \code{augment_region} returned by \code{get_augment_region}.}

\item{...}{Unused.}
}
\description{
Print method for augment_region objects
}
