% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.optdes}
\alias{print.optdes}
\title{Print function for optdes}
\usage{
\method{print}{optdes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{optdes}.}

\item{...}{Possible extra arguments for printing dataframes}
}
\description{
Print function for optdes
}
\examples{
rri <- opt_des(Criterion = "I-Optimality", model = y ~ a * exp(-b / x),
  parameters = c("a", "b"), par_values = c(1, 1500), design_space = c(212, 422),
  reg_int = c(380, 422))
print(rri)
}
