% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{sens_val_to_add}
\alias{sens_val_to_add}
\title{Calculates sensitivity function value for given delta and efficiency}
\usage{
sens_val_to_add(deff, alpha, k)
}
\arguments{
\item{deff}{Minimum efficiency of the resulting design}

\item{alpha}{Combined weight of the new points to add}

\item{k}{Number of parameters of the model}
}
\value{
Value of the sensitivity function over. Points with a sensitivity
function over that are suitable to be added.
}
\description{
Uses the formula to calculate the sensitivity function value that delimits
which points can be added to the design guaranteing the chosen efficiency.
}
\examples{
optedr:::sens_val_to_add(0.9, 1/4, 3)
}
