% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{shiny_augment}
\alias{shiny_augment}
\title{Shiny D-augment}
\usage{
shiny_augment()
}
\description{
Launches the demo Shiny application to D-augment several pre-specified
models. Requires an interactive R session; if called non-interactively
a message with the web URL is emitted instead.
}
\examples{
shiny_augment()
}
