% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{shiny_optimal}
\alias{shiny_optimal}
\title{Shiny Optimal}
\usage{
shiny_optimal()
}
\description{
Launches the demo Shiny application to calculate optimal designs for
Antoine's Equation. Requires an interactive R session; if called
non-interactively a message with the web URL is emitted instead.
}
\examples{
shiny_optimal()
}
