% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.optdes}
\alias{summary.optdes}
\title{Summary function for optdes}
\usage{
\method{summary}{optdes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{optdes}.}

\item{...}{Possible extra arguments for the summary}
}
\description{
Summary function for optdes
}
\examples{
\dontrun{
rri <- opt_des("I-Optimality", y ~ a * exp(-b / x), c("a", "b"), c(1, 1500), c(212, 422),
matB <- matrix(c(3, 1, 1, 2), nrow = 2))
summary(rri)}
}
