% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tr}
\alias{tr}
\title{Trace}
\usage{
tr(M)
}
\arguments{
\item{M}{The matrix from which to calculate the trace.}
}
\value{
The trace of the matrix.
}
\description{
Return the mathematical trace of a matrix, the sum of its diagonal elements.
}
\examples{
optedr:::tr(matrix(c(1, 0, 0, 1), nrow = 2))
}
