% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_design_total}
\alias{update_design_total}
\title{Merge close points of a design}
\usage{
update_design_total(design, delta)
}
\arguments{
\item{design}{The design to update. It's a dataframe with two columns:
\itemize{
\item \code{Point} contains the support points of the design.
\item \code{Weight} contains the corresponding weights of the \code{Point}s.
}}

\item{delta}{Threshold which defines how close two points have to be to any of the existing ones in order to
merge with them.}
}
\value{
The updated design.
}
\description{
Takes a design and merge together all points that are closer between them than a certain threshold \code{delta}.
}
\examples{
design <- data.frame("Point" = c(1, 5, 11, 12), "Weight" = rep(1 / 4, times = 4))
optedr:::update_design_total(design, 1.1)
}
