% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_function}
\alias{weight_function}
\title{Weight function per distribution}
\usage{
weight_function(model, char_vars, values, distribution = "Homoscedasticity")
}
\arguments{
\item{model}{formula describing the model to use. Must use x as the variable.}

\item{char_vars}{character vector with the parameters of the models, as written in the \code{formula}}

\item{values}{numeric vector with the parameters nominal values, in the same order as given in \code{parameters}.}

\item{distribution}{character variable specifying the probability distribution of the response. Can be one of the following:
\itemize{
\item 'Homoscedasticity'
\item 'Gamma', which can be used for exponential or normal heteroscedastic with constant relative error
\item 'Poisson'
\item 'Logistic'
\item 'Log-Normal' (work in progress)
}}
}
\value{
one variable function that represents the square of the structure of variance, in case of heteroscedastic variance of the response.
}
\description{
Weight function per distribution
}
