% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_mult.R
\name{WFMult}
\alias{WFMult}
\title{Master function for the cocktail algorithm, that calls the appropriate one given the criterion.}
\usage{
WFMult(
  init_design,
  grad,
  criterion,
  par_int = NA,
  matB = NA,
  design_space,
  grid.length,
  join_thresh,
  delete_thresh,
  k,
  delta_weights,
  tol,
  tol2,
  max_iter,
  compound = NULL,
  kl_spec = NULL
)
}
\arguments{
\item{init_design}{optional dataframe with the initial design for the algorithm.}

\item{grad}{function of partial derivatives of the model.}

\item{criterion}{character variable with the chosen optimality criterion.}

\item{par_int}{numeric vector with the index of the parameters of interest (Ds-Optimality).}

\item{matB}{optional matrix for L-optimality.}

\item{design_space}{named list with bounds for each design variable.}

\item{grid.length}{numeric value for the sensitivity search grid / LHS size.}

\item{join_thresh}{numeric value for the merge heuristic.}

\item{delete_thresh}{numeric value for the weight deletion threshold.}

\item{k}{number of unknown parameters of the model.}

\item{delta_weights}{numeric value in (0, 1), parameter of the algorithm.}

\item{tol}{numeric value for convergence of the weight loop.}

\item{tol2}{numeric value for the outer stop condition.}

\item{max_iter}{maximum number of outer iterations.}

\item{compound}{preprocessed list of compound criterion specifications (internal use).}

\item{kl_spec}{preprocessed list of KL-Optimality specifications (internal use).}
}
\value{
An object of class \code{optdes}.
}
\description{
Depending on the \code{criterion} the cocktail algorithm for the chosen criterion is called,
and the necessary parameters for the functions are given from the user input.
}
\seealso{
Other cocktail algorithms: 
\code{\link{CWFMult}()},
\code{\link{DWFMult}()},
\code{\link{DsWFMult}()},
\code{\link{IWFMult}()},
\code{\link{KLWFMult}()}
}
\concept{cocktail algorithms}
