% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{crosspoints}
\alias{crosspoints}
\title{Calculate crosspoints}
\usage{
crosspoints(val, sens, gridlength, tol, xmin, xmax)
}
\arguments{
\item{val}{Efficiency value to solve in the curve relationing the space of the design and efficiency of new design}

\item{sens}{Sensitivity function of the design for the model}

\item{gridlength}{Number of points in the grid to find the crosspoints}

\item{tol}{Tolerance that establishes how close two points close to one another are considered the same}

\item{xmin}{Minimum of the space of the design}

\item{xmax}{Maximum of the space of the design}
}
\value{
A numeric vector of crosspoints that define the candidate points region
}
\description{
Given the parameters for augmenting a design, this function calculates the
crosspoints in the efficiency function that delimit the candidate points
region
}
