% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_eff.R
\name{dscrit}
\alias{dscrit}
\title{Criterion function for Ds-Optimality}
\usage{
dscrit(M, par_int)
}
\arguments{
\item{M}{Information matrix for which the criterion value wants to be calculated.}

\item{par_int}{Numeric vector with the index of the parameters of interest of the model.}
}
\value{
Numeric value of the Ds-optimality criterion for the information matrix.
}
\description{
Calculates the value of the Ds-Optimality criterion, which follows the expression:
\deqn{\phi_D = \frac{|M_{22}|}{|M|}^{1/s}}
}
\examples{
optedr:::dscrit(matrix(c(1, 0.75, 0.75, 0.625), nrow = 2), c(2))
}
