% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findmax}
\alias{findmax}
\title{Find Maximum}
\usage{
findmax(sens, min, max, grid.length)
}
\arguments{
\item{sens}{A single variable numeric function to evaluate.}

\item{min}{Minimum value of the search interval.}

\item{max}{Maximum value of the search interval.}

\item{grid.length}{Length of the search interval.}
}
\value{
The value at which the maximum is obtained
}
\description{
Searches the maximum of a function over a grid on a given interval.
}
\examples{
optedr:::findmax(function(x) x^2 - x^3, 0, 1, 1000)
}
