% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findmaxval}
\alias{findmaxval}
\title{Find Maximum Value}
\usage{
findmaxval(sens, min, max, grid.length)
}
\arguments{
\item{sens}{A single variable numeric function to evaluate.}

\item{min}{Minimum value of the search interval.}

\item{max}{Maximum value of the search interval.}

\item{grid.length}{Length of the search interval.}
}
\value{
The value of the maximum
}
\description{
Searches the maximum of a function over a grid on a given interval.
}
