% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findminval}
\alias{findminval}
\title{Find Minimum Value}
\usage{
findminval(sens, min, max, grid.length)
}
\arguments{
\item{sens}{a single variable numeric function to evaluate.}

\item{min}{minimum value of the search grid.}

\item{max}{maximum value of the search grid.}

\item{grid.length}{length of the search grid.}
}
\value{
The value of the minimum
}
\description{
Searches the maximum of a function over a grid on a given grid.
}
