% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPar}
\alias{getPar}
\title{Parity of the crosspoints}
\usage{
getPar(cross)
}
\arguments{
\item{cross}{Vector of crosspoints in the sensitivity function given an efficiency and weight}
}
\value{
True if the number of crosspoints is even, false otherwise
}
\description{
Determines if the number of crosspoints is even or odd given the vector
of crosspoints
}
