% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_eff.R
\name{icrit}
\alias{icrit}
\title{Criterion function for I-Optimality}
\usage{
icrit(M, matB)
}
\arguments{
\item{M}{Information matrix for which the criterion value wants to be calculated.}

\item{matB}{Matrix of the integral of the information matrix over the interest region. Identity matrix for
A-Optimality.}
}
\value{
Numeric value of the Ds-optimality criterion for the information matrix.
}
\description{
Calculates the value of the Ds-Optimality criterion, which follows the expression:
\deqn{\phi_D = \frac{|M_{22}|}{|M|}^{1/s}}
}
