% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_mat_sens.R
\name{isens}
\alias{isens}
\title{Sensitivity function for I-Optimality}
\usage{
isens(grad, M, matB)
}
\arguments{
\item{grad}{A function in a single variable that returns the partial derivatives vector of the model.}

\item{M}{Information Matrix for the sensitivity function.}

\item{matB}{Matrix resulting from the integration of the one-point Information Matrix along the interest
region or lineal matrix for L-Optimality.}
}
\value{
The sensitivity function as a matrix of single variable.
}
\description{
Calculates the sensitivity function from the gradient vector, the Information Matrix and the integral of the
one-point Identity Matrix over the interest region. If instead the identity matrix is used, it can be used
for A-Optimality.
}
