% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.optdes}
\alias{plot.optdes}
\title{Plot function for optdes}
\usage{
\method{plot}{optdes}(x)
}
\arguments{
\item{x}{An object of class \code{optdes}.}

\item{...}{Possible extra arguments for plotting dataframes}
}
\description{
Plot function for optdes
}
\examples{
rri <- opt_des("I-Optimality", y ~ a * exp(-b / x), c("a", "b"), c(1, 1500), c(212, 422),
matB <- matrix(c(3, 1, 1, 2), nrow = 2))
plot(rri)
}
