% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.optdes}
\alias{print.optdes}
\title{Print function for optdes}
\usage{
\method{print}{optdes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{optdes}.}

\item{...}{Possible extra arguments for printing dataframes}
}
\description{
Print function for optdes
}
\examples{
rri <- opt_des("I-Optimality", y ~ a * exp(-b / x), c("a", "b"), c(1, 1500), c(212, 422), matB = matrix(c(3, 1, 1, 2), nrow = 2))
print(rri)
}
