% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{shiny_augment}
\alias{shiny_augment}
\title{Shiny D-augment}
\usage{
shiny_augment()
}
\description{
Launches the demo shiny application to D-augment several preespecified models
}
\examples{
shiny_augment()
}
