% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{shiny_optimal}
\alias{shiny_optimal}
\title{Shiny Optimal}
\usage{
shiny_optimal()
}
\description{
Launches the demo shiny application to calculate optimal designs for Antoine's Equation
}
\examples{
shiny_optimal()
}
