% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.optdes}
\alias{summary.optdes}
\title{Summary function for optdes}
\usage{
\method{summary}{optdes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{optdes}.}

\item{...}{Possible extra arguments for the summary}
}
\description{
Summary function for optdes
}
\examples{
rri <- opt_des(criterion = "I-Optimality", model = y ~ a * exp(-b / x),
  parameters = c("a", "b"), par_values = c(1, 1500), design_space = c(212, 422),
  reg_int = c(380, 422))
summary(rri)
}
