% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_design}
\alias{update_design}
\title{Update Design with new point}
\usage{
update_design(design, xmax, delta)
}
\arguments{
\item{design}{Design to update. It's a dataframe with two columns:
\itemize{
\item \code{Point} contains the support points of the design.
\item \code{Weight} contains the corresponding weights of the \code{Point}s.
}}

\item{xmax}{The point to add as a numeric value.}

\item{delta}{Threshold which defines how close the new point has to be to any of the existing ones in order to
merge with them.}
}
\value{
The updated design.
}
\description{
Updates a design adding a new point to it. If the added point is closer than \code{delta} to an existing
point of the design, the two points are merged together as their arithmetic average. Then updates the weights
to be equal to all points of the design.
}
\examples{
# Without merging:
design <- data.frame("Point" = c(1, 5, 9), "Weight" = rep(1 / 3, times = 3))
optedr:::update_design(design, 2, 0.5)

# Merging:
design <- data.frame("Point" = c(1, 5, 9), "Weight" = rep(1 / 3, times = 3))
optedr:::update_design(design, 2, 1.1)
}
